test(prompt2blog): split storage route coverage
